Distributed Constraints Satisfaction Project
In this project, we dealt with Distributed Constraints Satisfaction
problem.
The problem describes a group of agents, which each agent has a set of
values attached to and there are various constraints about their values
between any two agents.
The goal is to get a valid solution that satisfies the constraints –
value attached to any agent regarding its constraints with other agents.

For agents representation we've used infrastructure that simulates a
process as an agent and the communication between the agents is done by
sending messages between processes in order to create an environment in
which each process is independent, non-shared memory. The purpose of
sending messages is to inform the agent on the receiving end the partly
assignment values at that stage.

We have implemented three algorithms that can solve this problem:
• SBT - Synchronous Backtracking
• AFC - Asynchronous Forward Checking
• DisCBJ - Distributed Conflict-Directed Back Jumping

We performed a series of experiments comparing the algorithms' indices
taking into account the network load (the amount of messages sent) and
the number of tests for constraint between the agents and their values.
Combining these measures allowed us to test the effectiveness of any one
of the three algorithms, and comparing between them.